Calendar + Calendar2
By the end of class I was able to figure out how to properly switch
between months and clear out / redisplay in the div when changing
between months.  As far as changing the calendar everything is working
perfectly.  I still need to figure out what is causing january to come
back as undefined and why february is 1 year behind what it should be.
Because of these remaining issues I reverted Calendar.cshtml to the
static page to show your dad/ other people at the church if necessary.
The calendar with navigation is Calendar2.cshtml.  I will keep looking
in to getting january/february fixed the rest of the week.