Calendar Complete!
There was an issue where I was no recalculating the year before calling
other functions that caused january to have an off year when using the
back button, and february to always be behind.  January is now showing
up and building properly.  I was using an old variable rather than the
proper one.  Everything works fine now.  The only thing not "working" is
there are no EventTime's for the events.  I tried adding them to the
seed/initializer but nothing would change and i didnt want to risk
breaking it, so i'll talk to jordan about how he implimented that on
wednesday.